[REFACTOR] Revert to old file names (hopefully lowercase now)
