+++ src/react-app/components/ProteinNode.tsx
