+++ src/react-app/components/SimulationDrawer.tsx
