+++ src/react-app/components/edges/Functions.tsx
