+++ src/react-app/components/edges/MichaelisMenten.tsx
