+++ src/react-app/components/edges/ReversibleMassAction.tsx
